Updates for SC24 structure
