+++ Code_Exercises/Multi_Platform_With_SYCL/solution.cpp
@@ -1,0 +1,17 @@
+/*
+ SYCL Academy (c)
+
+ SYCL Academy is licensed under a Creative Commons
+ Attribution-ShareAlike 4.0 International License.
+
+ You should have received a copy of the license along with this
+ work.  If not, see <http://creativecommons.org/licenses/by-sa/4.0/>.
+*/
+
+//#define SYCL_ACADEMY_USING_COMPUTECPP
+
+#define CATCH_CONFIG_MAIN
+#include <catch2/catch.hpp>
+
+#include <sycl/sycl.hpp>
+
