--- Lesson_Materials/OneMKL_gemm/index.html
@@ -1,159 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-	<head>
-	    <meta charset="utf-8">
-		<link rel="stylesheet" href="../common-revealjs/css/reveal.css">
-		<link rel="stylesheet" href="../common-revealjs/css/theme/white.css">
-		<link rel="stylesheet" href="../common-revealjs/css/custom.css">
-		<script>
-			// This is needed when printing the slides to pdf
-			var link = document.createElement( 'link' );
-			link.rel = 'stylesheet';
-			link.type = 'text/css';
-			link.href = window.location.search.match( /print-pdf/gi ) ? '../common-revealjs/css/print/pdf.css' : '../common-revealjs/css/print/paper.css';
-			document.getElementsByTagName( 'head' )[0].appendChild( link );
-		</script>
-		<script>
-		    // This is used to display the static images on each slide,
-			// See global-images in this html file and custom.css
-			(function() {
-				if(window.addEventListener) {
-					window.addEventListener('load', () => {
-						let slides = document.getElementsByClassName("slide-background");
-
-						if (slides.length === 0) {
-							slides = document.getElementsByClassName("pdf-page")
-						}
-
-						// Insert global images on each slide
-						for(let i = 0, max = slides.length; i < max; i++) {
-							let cln = document.getElementById("global-images").cloneNode(true);
-							cln.removeAttribute("id");
-							slides[i].appendChild(cln);
-						}
-
-						// Remove top level global images
-						let elem = document.getElementById("global-images");
-						elem.parentElement.removeChild(elem);
-					}, false);
-				}
-			})();
-		</script>
-
-	</head>
-	<body>
-		<div class="reveal">
-			<div class="slides">
-			  <div id="global-images" class="global-images">
-			    <img src="../common-revealjs/images/sycl_academy.png" />
-			    <img src="../common-revealjs/images/sycl_logo.png" />
-			    <img src="../common-revealjs/images/trademarks.png" />
-			    <img src="../common-revealjs/images/codeplay.png" />
-			  </div>
-			  <!--Slide 1-->
-			  <section class="hbox">
-			    <div class="hbox" data-markdown>
-			      ## oneAPI Math Kernel Library (oneMKL) Interfaces
-			    </div>
-			  </section>
-			  <!--Slide 2-->
-			  <section class="hbox" data-markdown>
-			    ## Learning Objectives
-			    * Learn about oneMKL library, more specifically oneMKL Interfaces project
-			    * Learn about how to use GEMM APIs from oneMKL with both USM and buffer memory models
-			  </section>
-			  <!--Slide 3-->
-			  <section class="container" data-markdown>
-			    ## Resources
-			    * oneMKL Interfaces: https://github.com/oneapi-src/oneMKL
-			    * oneMKL specification: https://spec.oneapi.io/versions/latest/elements/oneMKL/source/index.html#
-			    * Important: What is the difference between the following oneMKL terms:
-			    (1) oneAPI Specification for oneMKL
-			    (2) oneAPI's oneMKL Interfaces Project
-			    (3) Intel(R) oneAPI's oneMKL Product
-			    https://github.com/oneapi-src/oneMKL?tab=readme-ov-file#onemkl
-			  </section>
-			  <!--Slide 4-->
-			  <section>
-			    <div class="hbox" data-markdown>
-			      #### Run-time dispatching
-			    </div>
-			    <div class="container">
-			      <div class="col">
-				<code class="code-100pc"><pre>
-#include &ltoneapi/mkl/blas.hpp&gt
-
-...
-
-sycl::queue cpu_queue(sycl::cpu_selector_v);
-sycl::queue gpu_queue(sycl::gpu_selector_v);
-
-using namespace oneapi::mkl;
-blas::column_major::gemm(cpu_queue, 
-                         transA, transB, m, ...);
-blas::column_major::gemm(gpu_queue, 
-                         transA, transB, m, ...);
-				</code></pre>
-			      </div>
-			      <div class="col" data-markdown>
-				* Backend is loaded at run-time based on device-vendor
-				* `$> icpx -fsycl –I$ONEMKL/include app.cpp`
-				* `$> icpx -fsycl app.o –L$ONEMKL/lib –lonemkl`
-			      </div>
-			    </div>
-			  </section>
-			  <!--Slide 5-->
-			  <section>
-			    <div class="hbox" data-markdown>
-			      #### Compile-time dispatching
-			    </div>
-			    <div class="container">
-			      <div class="col">
-				<code class="code-100pc"><pre>
-#include &ltoneapi/mkl/blas.hpp&gt
-
-...
-
-sycl::queue cpu_queue(sycl::cpu_selector_v);
-sycl::queue gpu_queue(sycl::gpu_selector_v);
-
-using namespace oneapi::mkl;
-backend_selector&ltbackend::mklcpu&gt cpu_selector(cpu_queue);
-backend_selector&ltbackend::cublas&gt gpu_selector(gpu_queue);
-
-blas::column_major::gemm(cpu_selector, 
-                         transA, transB, m, ...);
-blas::column_major::gemm(gpu_selector, 
-                         transA, transB, m, ...);
-				</code></pre>
-			      </div>
-			      <div class="container" data-markdown>
-				* Uses a templated backend selector APIs, where the template parameters specify the backends
-				* Application is linked with the required oneMKL backend wrapper libraries
-				* `$> clang++ -fsycl –I$ONEMKL/include app.cpp`
-				* `$> clang++ -fsycl app.o –L$ONEMKL/lib –lonemkl_blas_mklcpu –lonemkl_blas_cublas`
-			      </div>
-			    </div>
-			  </section>
-			  <!--Slide 6-->
-			  <section class="hbox" data-markdown>
-			    ## Exercise
-			    * Objectives: Learn to use oneMKL GEMM buffer, USM APIs
-			    * What is provided:
-			      * Boiler plate-code provided (a) to perform GEMM on CPU, (b) Helper function to verify results from oneMKL APIs and CPU
-			      * Please complete the TODO tasks marked in the `source_*.cpp`.
-			      * Refer to the solutions at `solution_*.cpp`
-			  </section>
-			</div>
-		</div>
-		<script src="../common-revealjs/js/reveal.js"></script>
-		<script src="../common-revealjs/plugin/markdown/marked.js"></script>
-		<script src="../common-revealjs/plugin/markdown/markdown.js"></script>
-		<script src="../common-revealjs/plugin/notes/notes.js"></script>
-		<script>
-			Reveal.initialize({mouseWheel: true, defaultNotes: true});
-			Reveal.configure({ slideNumber: true });
-		</script>
-	</body>
-</html>
